Remove Android.mk backend
