+++ Blueprints
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2020 Arm Limited.
+ * Copyright 2018-2020, 2022 Arm Limited.
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -39,7 +39,6 @@ bootstrap_go_package {
     ],
     srcs: [
         "core/android.go",
-        "core/android_make.go",
         "core/androidbp_backend.go",
         "core/androidbp_cclibs.go",
         "core/androidbp_kernel_module.go",
