+++ docs/config_system.md
@@ -395,5 +395,5 @@ Config plugins are included in the build system by adding them to the
 ```bash
 export BOB_CONFIG_PLUGINS="path/to/config_plugin:path/to/other_config_plugin" # No .py extension
 ... # Other exports for Bob bootstrap
-bob/bootstrap_linux.bash # or bootstrap_androidmk.bash
+bob/bootstrap_linux.bash # or bootstrap_androidbp.bash
 ```
