+++ example/build.bp
@@ -12,9 +12,6 @@ bob_defaults {
 
 bob_install_group {
     name: "bin",
-    builder_android_make: {
-        install_path: "$(TARGET_OUT)/bin",
-    },
     builder_android_bp: {
         install_path: "bin",
     },
