+++ mconfig/host_toolchain.Mconfig
@@ -1,4 +1,4 @@
-# Copyright 2016-2021 Arm Limited.
+# Copyright 2016-2022 Arm Limited.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -22,7 +22,7 @@
 
 config HOST_GNU_PREFIX
 	string "Host GNU compiler prefix"
-	default "prebuilts/gcc/linux-x86/x86/x86_64-linux-android-4.9/bin/x86_64-linux-android-" if BUILDER_ANDROID_MAKE || BUILDER_ANDROID_BP
+	default "prebuilts/gcc/linux-x86/x86/x86_64-linux-android-4.9/bin/x86_64-linux-android-" if BUILDER_ANDROID_BP
 	default ""
 
 config HOST_GNU_CC_BINARY
@@ -39,10 +39,8 @@ config HOST_GNU_CXX_BINARY
 
 config HOST_CLANG_PREFIX
 	string "Host Clang compiler prefix"
-	default "prebuilts/clang/host/linux-x86/clang-4691093/bin/" if ANDROID_PLATFORM_VERSION = 9 && (BUILDER_ANDROID_BP || BUILDER_ANDROID_MAKE) # pie-release
-	default "prebuilts/clang/host/linux-x86/clang-r353983c/bin/" if ANDROID_PLATFORM_VERSION = 10 && (BUILDER_ANDROID_BP || BUILDER_ANDROID_MAKE) # android-10-release
-	default "prebuilts/clang/host/linux-x86/clang-r383902b/bin/" if ANDROID_PLATFORM_VERSION = 11 && (BUILDER_ANDROID_BP || BUILDER_ANDROID_MAKE) # android-11-release
-	default "prebuilts/clang/host/linux-x86/clang-r416183b1/bin/" if ANDROID_PLATFORM_VERSION > 11 && (BUILDER_ANDROID_BP || BUILDER_ANDROID_MAKE)
+	default "prebuilts/clang/host/linux-x86/clang-r383902b/bin/" if ANDROID_PLATFORM_VERSION = 11 && BUILDER_ANDROID_BP # android-11-release
+	default "prebuilts/clang/host/linux-x86/clang-r416183b1/bin/" if ANDROID_PLATFORM_VERSION > 11 && BUILDER_ANDROID_BP
 	default ""
 	help
 	  This is typically the value of ClangDefaultVersion declared in <aosp>/soong/cc/config/global.go
