+++ tests/resources/build.bp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2020 Arm Limited.
+ * Copyright 2018-2020, 2022 Arm Limited.
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,9 +17,6 @@
 
 bob_install_group {
     name: "IG_testcases",
-    builder_android_make: {
-        install_path: "$(TARGET_OUT_TESTCASES)",
-    },
     builder_android_bp: {
         install_path: "tests",
     },
@@ -33,9 +30,6 @@ bob_resource {
     srcs: ["main.c"],
     install_group: "IG_testcases",
     build_by_default: true,
-    builder_android_make: {
-        relative_install_path: "x",
-    },
     builder_android_bp: {
         relative_install_path: "z",
     },
